team names render on the page
