+++ src/components/GameCard/GameCard.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, {useState} from "react";
 // import {Link} from 'react-router-dom' //TODO: add a link to team info
 import './GameCard.css'
